# ENH: re-implement SIMD kernels of complex operations