… benchmarks/benchmarks/bench_ufunc_strides.py
… numpy/core/src/common/simd/avx2/memory.h
… numpy/core/src/common/simd/avx512/memory.h
… numpy/core/src/common/simd/avx512/reorder.h
… numpy/core/src/common/simd/neon/memory.h
… numpy/core/src/common/simd/neon/reorder.h
… numpy/core/src/common/simd/sse/memory.h
… numpy/core/src/common/simd/sse/reorder.h
… numpy/core/src/common/simd/vec/memory.h
… numpy/core/src/common/simd/vec/reorder.h
… numpy/core/src/umath/loops.c.src
… numpy/core/src/umath/loops.h.src
… numpy/core/src/umath/loops_unary_complex.dispatch.c.src
… numpy/core/src/umath/simd.inc.src
… numpy/core/tests/test_simd.py