This pull-request:

- re-implement SIMD kernels of complex operations(multiply, add, sub, absolute, square, conjugate)

    New kernels implemented via universal intrinsic,
    which leads to bringing SIMD kernels for all supported SIMD extensions,
    also adds support for MSVC.

    The raw SIMD only supports AVX512,  still new kernels provide better
    performance for non-contiguous memory access and don't require 128-bit/64-bit alignment
    for complex128/complex64 just 64-bit/32-bit.

- Remove raw SIMD of f32/f64 arithmetic operations on X86 and
   activate the implementation of universal intrinsics instead.

- Implement intrinsics for mask division

- Implement intrinsics for shuffle over 128-bit lane and unzip
    
    shuffle intrinsics support 32-bit/64-bit vector data types,
    unzip(deinterleave) intrinsics supports all data types.

- Add special intrinsics for better non-contiguous/partial memory access for complex load/store
    
    summarized as follows:
    
    * 64-bit contiguous partial load/store over 32-bit lane
       npyv_load2_till_u32, npyv_load2_till_s32, npyv_load2_till_f32
       npyv_load2_tillz_u32, npyv_load2_tillz_s32, npyv_load2_tillz_f32
       npyv_store2_till_u32, npyv_store2_till_s32, npyv_store2_till_f32
    
    * 128-bit contiguous partial load/store over 64-bit lane
       npyv_load2_till_u64, npyv_load2_till_s64, npyv_load2_till_f64
       npyv_load2_tillz_u64, npyv_load2_tillz_s64, npyv_load2_tillz_f64
       npyv_store2_till_u64, npyv_store2_till_s64, npyv_store2_till_f64
    
    * 64-bit non-contiguous load/store over 32-bit stride
       npyv_loadn2_u32,  npyv_loadn2_s32,  npyv_loadn2_f32
       npyv_storen2_u32, npyv_storen2_s32, npyv_storen2_f32
    
    * 128-bit non-contiguous load/store over 64-bit stride
       npyv_loadn2_u64,  npyv_loadn2_s64,  npyv_loadn2_f64
       npyv_storen2_u64, npyv_storen2_s64, npyv_storen2_f64
    
    * 64-bit non-contiguous partial load/store over 32-bit stride
       npyv_loadn2_till_u32, npyv_loadn2_till_s32, npyv_loadn2_till_f32
       npyv_loadn2_tillz_u32, npyv_loadn2_tillz_s32, npyv_loadn2_tillz_f32
       npyv_storen2_till_u32, npyv_storen2_till_s32, npyv_storen2_till_f32
    
    * 128-bit non-contiguous partial load/store over 64-bit stride
       npyv_loadn2_till_u64, npyv_loadn2_till_s64, npyv_loadn2_till_f64
       npyv_loadn2_tillz_u64, npyv_loadn2_tillz_s64, npyv_loadn2_tillz_f64
       npyv_storen2_till_u64, npyv_storen2_till_s64, npyv_storen2_till_f64
    
    * 2 channels de-interlave/interleave contiguous load/store for all data types
          npyv_load_##sfx##x2, npyv_store_##sfx##x2

------

### X86

<details>
<summary>CPU</summary>

```Bash
Architecture:            x86_64
  CPU op-mode(s):        32-bit, 64-bit
  Address sizes:         48 bits physical, 48 bits virtual
  Byte Order:            Little Endian
CPU(s):                  16
  On-line CPU(s) list:   0-15
Vendor ID:               AuthenticAMD
  Model name:            AMD Ryzen 7 7700X 8-Core Processor
    CPU family:          25
    Model:               97
    Thread(s) per core:  2
    Core(s) per socket:  8
    Socket(s):           1
    Stepping:            2
    Frequency boost:     disabled
    CPU(s) scaling MHz:  67%
    CPU max MHz:         5572.2651
    CPU min MHz:         3000.0000
    BogoMIPS:            8986.45
    Flags:               fpu vme de pse tsc msr pae mce cx8 apic sep
                         mtrr pge mca cmov pat pse36 clflush mmx fxsr
                          sse sse2 ht syscall nx mmxext fxsr_opt pdpe
                         1gb rdtscp lm constant_tsc rep_good amd_lbr_
                         v2 nopl nonstop_tsc cpuid extd_apicid aperfm
                         perf rapl pni pclmulqdq monitor ssse3 fma cx
                         16 sse4_1 sse4_2 x2apic movbe popcnt aes xsa
                         ve avx f16c rdrand lahf_lm cmp_legacy svm ex
                         tapic cr8_legacy abm sse4a misalignsse 3dnow
                         prefetch osvw ibs skinit wdt tce topoext per
                         fctr_core perfctr_nb bpext perfctr_llc mwait
                         x cpb cat_l3 cdp_l3 hw_pstate ssbd mba perfm
                         on_v2 ibrs ibpb stibp vmmcall fsgsbase bmi1
                         avx2 smep bmi2 erms invpcid cqm rdt_a avx512
                         f avx512dq rdseed adx smap avx512ifma clflus
                         hopt clwb avx512cd sha_ni avx512bw avx512vl
                         xsaveopt xsavec xgetbv1 xsaves cqm_llc cqm_o
                         ccup_llc cqm_mbm_total cqm_mbm_local avx512_
                         bf16 clzero irperf xsaveerptr rdpru wbnoinvd
                          cppc arat npt lbrv svm_lock nrip_save tsc_s
                         cale vmcb_clean flushbyasid decodeassists pa
                         usefilter pfthreshold avic v_vmsave_vmload v
                         gif x2avic v_spec_ctrl avx512vbmi umip pku o
                         spke avx512_vbmi2 gfni vaes vpclmulqdq avx51
                         2_vnni avx512_bitalg avx512_vpopcntdq rdpid
                         overflow_recov succor smca fsrm flush_l1d
Virtualization features:
  Virtualization:        AMD-V
Caches (sum of all):
  L1d:                   256 KiB (8 instances)
  L1i:                   256 KiB (8 instances)
  L2:                    8 MiB (8 instances)
  L3:                    32 MiB (1 instance)
NUMA:
  NUMA node(s):          1
  NUMA node0 CPU(s):     0-15
Vulnerabilities:
  Itlb multihit:         Not affected
  L1tf:                  Not affected
  Mds:                   Not affected
  Meltdown:              Not affected
  Mmio stale data:       Not affected
  Retbleed:              Not affected
  Spec store bypass:     Mitigation; Speculative Store Bypass disable
                         d via prctl
  Spectre v1:            Mitigation; usercopy/swapgs barriers and __u
                         ser pointer sanitization
  Spectre v2:            Mitigation; Retpolines, IBPB conditional, IB
                         RS_FW, STIBP always-on, RSB filling, PBRSB-e
                         IBRS Not affected
  Srbds:                 Not affected
  Tsx async abort:       Not affected
```
</details>

<details>
<summary>OS</summary>

```Bash
Linux seiko-pc 6.1.1-arch1-1 #1 SMP PREEMPT_DYNAMIC Wed, 21 Dec 2022 22:27:55 +0000 x86_64 GNU/Linux
Python 3.10.9
gcc (GCC) 12.2.0
```
</details>

#### Benchmark
<details>
 <summary>AVX512F(before) vs AVX2(except for absolute)</summary>

```Bash
unset NPY_DISABLE_CPU_FEATURES
python runtests.py -n --bench-compare parent/main Complex  -- --cpu-affinity=6,7
```
```Bash
       before           after         ratio
     [6d474f2d]       [62fef37f]
                      <replace_raw_arithmfp>
-     4.79±0.03μs      4.55±0.05μs     0.95  bench_ufunc_strides.BinaryComplex.time_ufunc_scalar_in1('subtract', 4, 4, 2, 'F')
-      4.86±0.1μs      4.61±0.06μs     0.95  bench_ufunc_strides.BinaryComplex.time_ufunc_scalar_in0('add', 4, 4, 2, 'F')
-      4.85±0.1μs      4.60±0.05μs     0.95  bench_ufunc_strides.BinaryComplex.time_ufunc_scalar_in0('add', 4, 2, 4, 'F')
-     4.41±0.02μs      4.18±0.08μs     0.95  bench_ufunc_strides.BinaryComplex.time_ufunc('add', 2, 2, 2, 'F')
-     5.23±0.04μs      4.96±0.05μs     0.95  bench_ufunc_strides.BinaryComplex.time_ufunc('multiply', 1, 1, 4, 'F')
-      6.40±0.1μs      6.05±0.02μs     0.95  bench_ufunc_strides.BinaryComplex.time_ufunc_scalar_in1('multiply', 2, 2, 2, 'D')
-     4.79±0.05μs      4.53±0.05μs     0.94  bench_ufunc_strides.BinaryComplex.time_ufunc_scalar_in1('add', 4, 4, 2, 'F')
-     6.47±0.03μs      6.11±0.06μs     0.94  bench_ufunc_strides.BinaryComplex.time_ufunc_scalar_in1('multiply', 2, 1, 2, 'D')
-     4.79±0.04μs      4.52±0.08μs     0.94  bench_ufunc_strides.BinaryComplex.time_ufunc_scalar_in1('add', 2, 1, 4, 'F')
-     4.44±0.05μs      4.18±0.04μs     0.94  bench_ufunc_strides.BinaryComplex.time_ufunc('subtract', 1, 1, 4, 'F')
-     4.75±0.06μs      4.47±0.03μs     0.94  bench_ufunc_strides.BinaryComplex.time_ufunc_scalar_in0('subtract', 4, 4, 2, 'F')
-      7.75±0.2μs      7.28±0.06μs     0.94  bench_ufunc_strides.UnaryComplex.time_ufunc('square', 4, 1, 'D')
-     4.88±0.06μs       4.59±0.1μs     0.94  bench_ufunc_strides.BinaryComplex.time_ufunc_scalar_in1('add', 2, 2, 4, 'F')
-      6.50±0.1μs      6.10±0.03μs     0.94  bench_ufunc_strides.BinaryComplex.time_ufunc_scalar_in1('multiply', 4, 2, 4, 'F')
-     5.12±0.09μs      4.78±0.02μs     0.94  bench_ufunc_strides.BinaryComplex.time_ufunc('add', 4, 1, 2, 'F')
-     5.04±0.02μs      4.70±0.07μs     0.93  bench_ufunc_strides.BinaryComplex.time_ufunc('multiply', 1, 2, 2, 'F')
-     6.51±0.08μs      6.07±0.04μs     0.93  bench_ufunc_strides.BinaryComplex.time_ufunc_scalar_in0('multiply', 1, 2, 2, 'D')
-      6.46±0.1μs       6.01±0.1μs     0.93  bench_ufunc_strides.BinaryComplex.time_ufunc_scalar_in1('multiply', 4, 1, 4, 'F')
-      4.90±0.1μs      4.56±0.04μs     0.93  bench_ufunc_strides.BinaryComplex.time_ufunc_scalar_in1('subtract', 2, 1, 4, 'F')
-      6.43±0.1μs      5.97±0.08μs     0.93  bench_ufunc_strides.BinaryComplex.time_ufunc_scalar_in0('multiply', 2, 2, 2, 'D')
-      6.50±0.1μs      6.03±0.09μs     0.93  bench_ufunc_strides.BinaryComplex.time_ufunc_scalar_in1('multiply', 4, 4, 4, 'F')
-     4.41±0.05μs      4.09±0.08μs     0.93  bench_ufunc_strides.BinaryComplex.time_ufunc('subtract', 1, 4, 1, 'F')
-     4.52±0.09μs      4.18±0.05μs     0.93  bench_ufunc_strides.BinaryComplex.time_ufunc('subtract', 2, 2, 2, 'F')
-     14.0±0.04μs      13.0±0.04μs     0.93  bench_ufunc_strides.UnaryComplex.time_ufunc('absolute', 4, 2, 'F')
-     4.87±0.07μs      4.50±0.08μs     0.92  bench_ufunc_strides.BinaryComplex.time_ufunc_scalar_in0('add', 1, 4, 2, 'F')
-     13.8±0.02μs      12.7±0.02μs     0.92  bench_ufunc_strides.UnaryComplex.time_ufunc('absolute', 2, 2, 'F')
-     5.07±0.03μs      4.68±0.01μs     0.92  bench_ufunc_strides.BinaryComplex.time_ufunc('multiply', 2, 1, 2, 'F')
-     4.87±0.08μs       4.48±0.1μs     0.92  bench_ufunc_strides.BinaryComplex.time_ufunc_scalar_in0('subtract', 2, 4, 2, 'F')
-     4.46±0.04μs      4.10±0.05μs     0.92  bench_ufunc_strides.BinaryComplex.time_ufunc('add', 1, 1, 4, 'F')
-     11.4±0.06μs      10.5±0.03μs     0.92  bench_ufunc_strides.UnaryComplex.time_ufunc('absolute', 4, 1, 'F')
-     5.03±0.02μs      4.62±0.08μs     0.92  bench_ufunc_strides.BinaryComplex.time_ufunc('multiply', 1, 2, 1, 'F')
-     11.2±0.01μs      10.3±0.01μs     0.92  bench_ufunc_strides.UnaryComplex.time_ufunc('absolute', 2, 1, 'F')
-     5.18±0.04μs      4.75±0.06μs     0.92  bench_ufunc_strides.BinaryComplex.time_ufunc('multiply', 1, 4, 1, 'F')
-     5.07±0.01μs      4.65±0.03μs     0.92  bench_ufunc_strides.BinaryComplex.time_ufunc('multiply', 2, 2, 1, 'F')
-      14.0±0.1μs      12.8±0.07μs     0.92  bench_ufunc_strides.UnaryComplex.time_ufunc('absolute', 2, 4, 'F')
-     14.5±0.03μs      13.3±0.06μs     0.92  bench_ufunc_strides.UnaryComplex.time_ufunc('absolute', 4, 4, 'F')
-     5.04±0.02μs      4.62±0.05μs     0.92  bench_ufunc_strides.BinaryComplex.time_ufunc('multiply', 2, 1, 1, 'F')
-     5.06±0.03μs      4.63±0.02μs     0.91  bench_ufunc_strides.BinaryComplex.time_ufunc('multiply', 1, 1, 2, 'F')
-      4.85±0.2μs      4.41±0.01μs     0.91  bench_ufunc_strides.BinaryComplex.time_ufunc_scalar_in0('subtract', 1, 4, 2, 'F')
-      4.99±0.1μs      4.54±0.05μs     0.91  bench_ufunc_strides.BinaryComplex.time_ufunc_scalar_in1('subtract', 4, 1, 2, 'F')
-     4.50±0.07μs      4.09±0.03μs     0.91  bench_ufunc_strides.BinaryComplex.time_ufunc('add', 4, 1, 1, 'F')
-     5.30±0.04μs      4.81±0.06μs     0.91  bench_ufunc_strides.BinaryComplex.time_ufunc('multiply', 4, 1, 1, 'F')
-     5.60±0.03μs       5.09±0.1μs     0.91  bench_ufunc_strides.BinaryComplex.time_ufunc('multiply', 4, 1, 2, 'F')
-     5.20±0.03μs      4.72±0.06μs     0.91  bench_ufunc_strides.BinaryComplex.time_ufunc_scalar_in0('add', 1, 1, 2, 'D')
-     5.17±0.04μs      4.69±0.09μs     0.91  bench_ufunc_strides.BinaryComplex.time_ufunc_scalar_in1('add', 1, 4, 2, 'D')
-       138±0.5μs        125±0.8μs     0.91  bench_core.VarComplex.time_var(100000)
-      4.51±0.1μs      4.08±0.02μs     0.90  bench_ufunc_strides.BinaryComplex.time_ufunc('subtract', 4, 1, 1, 'F')
-     5.19±0.07μs      4.68±0.07μs     0.90  bench_ufunc_strides.BinaryComplex.time_ufunc_scalar_in0('add', 4, 2, 1, 'D')
-     5.17±0.02μs      4.66±0.03μs     0.90  bench_ufunc_strides.BinaryComplex.time_ufunc_scalar_in0('subtract', 4, 1, 2, 'D')
-     5.23±0.01μs      4.71±0.08μs     0.90  bench_ufunc_strides.BinaryComplex.time_ufunc_scalar_in1('subtract', 1, 4, 2, 'D')
-     5.23±0.03μs      4.71±0.08μs     0.90  bench_ufunc_strides.BinaryComplex.time_ufunc_scalar_in0('subtract', 1, 1, 2, 'D')
-     4.00±0.01μs         3.58±0μs     0.90  bench_ufunc_strides.UnaryComplex.time_ufunc('conjugate', 2, 2, 'F')
-     5.23±0.04μs      4.69±0.07μs     0.90  bench_ufunc_strides.BinaryComplex.time_ufunc_scalar_in1('add', 2, 2, 1, 'D')
-     5.20±0.03μs       4.65±0.1μs     0.89  bench_ufunc_strides.BinaryComplex.time_ufunc_scalar_in0('add', 1, 2, 1, 'D')
-     5.20±0.02μs       4.64±0.2μs     0.89  bench_ufunc_strides.BinaryComplex.time_ufunc_scalar_in1('subtract', 2, 4, 1, 'D')
-     5.22±0.02μs       4.66±0.1μs     0.89  bench_ufunc_strides.BinaryComplex.time_ufunc_scalar_in0('subtract', 2, 1, 2, 'D')
-     5.17±0.07μs       4.61±0.1μs     0.89  bench_ufunc_strides.BinaryComplex.time_ufunc_scalar_in1('add', 2, 4, 1, 'D')
-     5.22±0.04μs       4.65±0.1μs     0.89  bench_ufunc_strides.BinaryComplex.time_ufunc_scalar_in0('add', 2, 1, 2, 'D')
-     5.74±0.05μs      5.09±0.08μs     0.89  bench_ufunc_strides.BinaryComplex.time_ufunc('multiply', 4, 2, 1, 'F')
-     5.24±0.06μs       4.63±0.1μs     0.88  bench_ufunc_strides.BinaryComplex.time_ufunc_scalar_in0('subtract', 2, 2, 1, 'D')
-     5.24±0.03μs      4.63±0.02μs     0.88  bench_ufunc_strides.BinaryComplex.time_ufunc_scalar_in1('subtract', 1, 1, 2, 'D')
-     5.23±0.02μs      4.61±0.09μs     0.88  bench_ufunc_strides.BinaryComplex.time_ufunc_scalar_in1('add', 1, 1, 2, 'D')
-     5.22±0.01μs      4.58±0.03μs     0.88  bench_ufunc_strides.BinaryComplex.time_ufunc_scalar_in0('add', 4, 1, 2, 'D')
-     5.24±0.04μs      4.59±0.04μs     0.88  bench_ufunc_strides.BinaryComplex.time_ufunc_scalar_in1('subtract', 2, 1, 1, 'D')
-     5.21±0.04μs      4.55±0.06μs     0.87  bench_ufunc_strides.BinaryComplex.time_ufunc_scalar_in1('add', 2, 1, 1, 'D')
-     5.24±0.02μs      4.57±0.07μs     0.87  bench_ufunc_strides.BinaryComplex.time_ufunc_scalar_in1('add', 1, 2, 2, 'D')
-     5.27±0.03μs      4.60±0.09μs     0.87  bench_ufunc_strides.BinaryComplex.time_ufunc_scalar_in0('add', 2, 2, 1, 'D')
-     5.21±0.02μs      4.54±0.07μs     0.87  bench_ufunc_strides.BinaryComplex.time_ufunc_scalar_in1('subtract', 1, 2, 2, 'D')
-      8.14±0.1μs      7.09±0.04μs     0.87  bench_ufunc_strides.UnaryComplex.time_ufunc('conjugate', 4, 1, 'D')
-     5.21±0.03μs      4.51±0.03μs     0.87  bench_ufunc_strides.BinaryComplex.time_ufunc_scalar_in1('subtract', 2, 2, 1, 'D')
-     5.23±0.04μs      4.52±0.06μs     0.86  bench_ufunc_strides.BinaryComplex.time_ufunc_scalar_in0('subtract', 4, 2, 1, 'D')
-     5.22±0.03μs      4.50±0.04μs     0.86  bench_ufunc_strides.BinaryComplex.time_ufunc_scalar_in0('subtract', 1, 2, 1, 'D')
-     4.60±0.04μs      3.96±0.07μs     0.86  bench_ufunc_strides.BinaryComplex.time_ufunc_scalar_in1('subtract', 1, 1, 4, 'F')
-     4.04±0.03μs      3.45±0.05μs     0.85  bench_ufunc_strides.UnaryComplex.time_ufunc('conjugate', 1, 4, 'F')
-     4.62±0.02μs      3.93±0.07μs     0.85  bench_ufunc_strides.BinaryComplex.time_ufunc_scalar_in0('subtract', 1, 1, 4, 'F')
-     4.64±0.04μs      3.95±0.08μs     0.85  bench_ufunc_strides.BinaryComplex.time_ufunc_scalar_in1('subtract', 1, 4, 4, 'F')
-     4.60±0.03μs      3.91±0.02μs     0.85  bench_ufunc_strides.BinaryComplex.time_ufunc_scalar_in1('subtract', 4, 1, 1, 'F')
-     4.22±0.04μs      3.59±0.06μs     0.85  bench_ufunc_strides.BinaryComplex.time_ufunc('add', 2, 2, 1, 'F')
-     4.58±0.05μs      3.87±0.03μs     0.85  bench_ufunc_strides.BinaryComplex.time_ufunc_scalar_in1('subtract', 4, 2, 1, 'F')
-     4.23±0.06μs      3.58±0.04μs     0.85  bench_ufunc_strides.BinaryComplex.time_ufunc('subtract', 1, 2, 2, 'F')
-     4.63±0.06μs      3.91±0.03μs     0.84  bench_ufunc_strides.BinaryComplex.time_ufunc_scalar_in0('subtract', 2, 1, 4, 'F')
-     4.67±0.05μs      3.93±0.06μs     0.84  bench_ufunc_strides.BinaryComplex.time_ufunc_scalar_in0('subtract', 4, 1, 4, 'F')
-     5.47±0.05μs      4.60±0.07μs     0.84  bench_ufunc_strides.BinaryComplex.time_ufunc_scalar_in0('multiply', 4, 2, 4, 'F')
-     4.63±0.08μs      3.89±0.05μs     0.84  bench_ufunc_strides.BinaryComplex.time_ufunc_scalar_in1('add', 1, 4, 4, 'F')
-     4.63±0.01μs      3.89±0.03μs     0.84  bench_ufunc_strides.BinaryComplex.time_ufunc_scalar_in0('add', 2, 1, 4, 'F')
-     4.61±0.03μs      3.87±0.01μs     0.84  bench_ufunc_strides.BinaryComplex.time_ufunc_scalar_in0('add', 1, 1, 4, 'F')
-     4.63±0.04μs      3.88±0.04μs     0.84  bench_ufunc_strides.BinaryComplex.time_ufunc_scalar_in0('add', 4, 1, 4, 'F')
-     4.65±0.04μs      3.90±0.09μs     0.84  bench_ufunc_strides.BinaryComplex.time_ufunc_scalar_in0('add', 4, 4, 1, 'F')
-     4.62±0.04μs      3.86±0.02μs     0.84  bench_ufunc_strides.BinaryComplex.time_ufunc_scalar_in0('subtract', 1, 4, 1, 'F')
-     4.64±0.05μs      3.88±0.04μs     0.84  bench_ufunc_strides.BinaryComplex.time_ufunc_scalar_in1('add', 4, 4, 1, 'F')
-      23.4±0.2μs      19.6±0.07μs     0.84  bench_ufunc_strides.UnaryComplex.time_ufunc('absolute', 2, 4, 'D')
-     4.64±0.03μs      3.88±0.04μs     0.84  bench_ufunc_strides.BinaryComplex.time_ufunc_scalar_in1('subtract', 1, 2, 4, 'F')
-     4.62±0.06μs      3.86±0.04μs     0.84  bench_ufunc_strides.BinaryComplex.time_ufunc_scalar_in1('add', 4, 2, 1, 'F')
-     4.60±0.04μs      3.84±0.04μs     0.83  bench_ufunc_strides.BinaryComplex.time_ufunc_scalar_in0('subtract', 4, 4, 1, 'F')
-     4.63±0.03μs      3.86±0.06μs     0.83  bench_ufunc_strides.BinaryComplex.time_ufunc_scalar_in1('add', 1, 2, 4, 'F')
-     4.66±0.05μs      3.88±0.07μs     0.83  bench_ufunc_strides.BinaryComplex.time_ufunc_scalar_in1('add', 1, 1, 4, 'F')
-     4.61±0.06μs      3.84±0.01μs     0.83  bench_ufunc_strides.BinaryComplex.time_ufunc_scalar_in0('subtract', 2, 4, 1, 'F')
-     4.65±0.04μs      3.87±0.02μs     0.83  bench_ufunc_strides.BinaryComplex.time_ufunc_scalar_in1('add', 4, 1, 1, 'F')
-     4.63±0.03μs      3.86±0.05μs     0.83  bench_ufunc_strides.BinaryComplex.time_ufunc_scalar_in0('add', 2, 4, 1, 'F')
-     4.64±0.04μs       3.86±0.1μs     0.83  bench_ufunc_strides.BinaryComplex.time_ufunc_scalar_in1('subtract', 4, 4, 1, 'F')
-     3.58±0.01μs      2.97±0.06μs     0.83  bench_ufunc_strides.UnaryComplex.time_ufunc('square', 1, 1, 'D')
-     18.2±0.03μs      15.1±0.08μs     0.83  bench_ufunc_strides.UnaryComplex.time_ufunc('absolute', 2, 1, 'D')
-     4.21±0.02μs      3.49±0.02μs     0.83  bench_ufunc_strides.BinaryComplex.time_ufunc('subtract', 2, 1, 1, 'F')
-     4.27±0.09μs      3.53±0.03μs     0.83  bench_ufunc_strides.BinaryComplex.time_ufunc('subtract', 2, 1, 2, 'F')
-     20.2±0.09μs      16.7±0.08μs     0.83  bench_ufunc_strides.UnaryComplex.time_ufunc('absolute', 2, 2, 'D')
-     4.20±0.02μs      3.47±0.01μs     0.83  bench_ufunc_strides.BinaryComplex.time_ufunc('subtract', 1, 2, 1, 'F')
-     5.56±0.03μs      4.60±0.06μs     0.83  bench_ufunc_strides.BinaryComplex.time_ufunc_scalar_in1('multiply', 2, 4, 4, 'F')
…